LayoutTests/crypto/subtle/X25519-generate-key-derive-bits.html

<!DOCTYPE html>
<html>
<head>
<script src="../../resources/js-test-pre.js"></script>
<script src="../resources/common.js"></script>
</head>
<body>
<p id="description"></p>
<div id="console"></div>

<script>
description("Test X25519 deriveBits operation with generated base key and null length");

jsTestIsAsync = true;

var extractable = true;
var rawPublicKey = new Uint8Array([233, 47, 177, 106, 82, 224, 122, 208, 149, 79, 30, 251, 246, 35, 88, 194, 200, 75, 31, 47, 233, 93, 241, 52, 51, 218, 46, 210, 51, 54, 156, 4]);


crypto.subtle.generateKey({ name: "X25519" }, extractable, ["deriveBits"]).then(function(result) {
    privateKey = result.privateKey;
    return crypto.subtle.importKey("raw", rawPublicKey, { name: "X25519" }, extractable, [ ]);
}).then(function(result) {
    publicKey = result;
    return crypto.subtle.deriveBits({ name:"X25519", public:publicKey }, privateKey, null);
}).then(function(result) {
    derivedKey = result;

    shouldBe("derivedKey.byteLength", "32");

    finishJSTest();
}, failAndFinishJSTest);

</script>

<script src="../../resources/js-test-post.js"></script>
</body>
</html>
